+++ api/package.json
@@ -7,7 +7,7 @@
   "license": "UNLICENSED",
   "scripts": {
     "codegen": "kysely-codegen --out-file ./src/types/schema.ts --dialect postgres",
-    "migrate": "ts-node src/migrations/runMigration.ts",
+    "migrate": "npx ts-node src/migrations/runMigration.ts",
     "seed": "ts-node src/migrations/seed.ts",
     "build": "nest build",
     "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
@@ -50,8 +50,7 @@
     "pg": "^8.11.3",
     "reflect-metadata": "^0.1.13",
     "rxjs": "^7.8.1",
-    "socket.io": "^4.7.4",
-    "swagger": "^0.7.5"
+    "socket.io": "^4.7.4"
   },
   "devDependencies": {
     "@nestjs/cli": "^10.3.0",
