+++ api/src/pong/pong.gateway.ts
@@ -55,7 +55,6 @@ const getPlayer = (
 };
 
 @WebSocketGateway(5050, {
-  // namespace: 'games',
   cors: {
     origin: '*',
   },
