+++ client/Dockerfile
@@ -6,4 +6,4 @@ COPY    package*.json ./
 
 RUN     npm i
 
-CMD     npm run migrate && npm run start:dev
+CMD     npm run dev
