+++ client/package.json
@@ -15,7 +15,6 @@
     "@mui/icons-material": "^5.14.19",
     "@mui/material": "^5.14.19",
     "@tanstack/react-query": "^5.17.9",
-    "@types/socket.io-client": "^3.0.0",
     "axios": "^1.6.2",
     "dotenv": "^16.3.1",
     "luxon": "^3.4.4",
