+++ client/vite.config.ts
@@ -11,7 +11,7 @@ export default defineConfig({
       "/api": process.env.REACT_APP_API_HOST,
       "/public": process.env.REACT_APP_API_HOST,
       "/socket.io/": {
-        target: "http://0.0.0.0:5050/",
+        target: process.env.GAME_SOCKET_SERVER_HOST,
         ws: true,
         // changeOrigin: true,
         // rewrite: (path) => path.replace(/^\/socket.io/, ""),
