+++ docker-compose.yml
@@ -31,17 +31,23 @@ services:
       - /app/node_modules
     ports:
       - 5000:5000
+    command: sh -c "npm run migrate && npm run start:dev"
     restart: always
     stop_grace_period: 1s
 
   client:
     container_name: frontend
-    build: ./
+    depends_on:
+      - api
+    build: ./client
     environment:
       - REACT_APP_API_HOST
+      - GAME_SOCKET_SERVER_HOST=http://backend:5050
     volumes:
       - ./client/:/app
       - /app/node_modules
+      - ./api/src/types:/api/src/types
+      - ./api/src/pong/gameLogic/gamePositions.ts:/api/src/pong/gameLogic/gamePositions.ts
     ports:
       - 3000:3000
     restart: always
